src/diffusers/examples/instruct_pix2pix/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
actionlib==1.14.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
angles==1.9.13
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
base_local_planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.8
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
datasets==3.5.0
Deprecated==1.2.18
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
diffusers @ file:///home/sauravdosi/mediffuse/diffusers
dill==0.3.6
distlib==0.3.8
dynamic-reconfigure==1.7.3
einops==0.8.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
git-lfs==1.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
hjson==3.1.0
huggingface-hub==0.30.2
humanize==4.12.2
idna==3.10
image-geometry==1.16.2
imageio==2.37.0
importlib_metadata==8.6.1
interactive-markers==1.12.0
Jinja2==3.1.6
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==3.9.2
laser_geometry==1.6.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
message-filters==1.17.0
ml_dtypes==0.5.1
monai==1.4.0
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.14
namex==0.0.9
networkx==3.4.2
nibabel==5.2.0
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
peft==0.15.2
pillow==10.2.0
pipenv==2024.1.0
piqa==1.3.2
platformdirs==4.3.6
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2024.11.6
requests==2.32.3
resource_retriever==1.12.7
responses==0.18.0
rich==14.0.0
rosapi==0.11.17
rosbag==1.17.0
rosboost-cfg==1.15.8
rosbridge-library==0.11.17
rosbridge-server==0.11.17
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
safetensors==0.5.3
scikit-image==0.22.0
scipy==1.15.2
sensor-msgs==1.13.1
sentencepiece==0.2.0
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
srdfdom==0.6.4
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf2-sensor-msgs==0.7.7
tifffile==2025.3.30
tokenizers==0.21.1
topic-tools==1.17.0
torch==2.6.0
torchio==0.20.7
torchvision==0.21.0
tqdm==4.67.1
trac_ik_python==1.6.6
transformers==4.51.3
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urdfdom-py==0.4.6
urllib3==2.4.0
virtualenv==20.26.6
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xacro==1.14.18
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
